solution/notifier/requirements.txt
redis==5.0.1
python-dotenv==1.0.0
jinja2==3.1.2
resend==0.8.0